requirements.txt
requests>=2.32.5
httpx>=0.27.0
python-dotenv==1.0.0
aiohttp==3.9.0
